Length and weight ranges for relevant species

Data frame containing information about length and individual weight ranges for relevant species, including quantiles and various statistics. It also contains metadata about faunistic category and period of relevance as target species.

data("DataTargetSpecies")

Format

A data frame with 392 observations on the following 26 variables:

SPECIES

Rubincode species identifier (character).

WMIN

Minimum observed individual weight (numeric).

WMIN05

5th percentile of individual weight (numeric).

WMIN10

10th percentile of individual weight (numeric).

WMIN25

25th percentile of individual weight (numeric).

WMAX_75

75th percentile of individual weight (numeric).

WMAX90

90th percentile of individual weight (numeric).

WMAX95

95th percentile of individual weight (numeric).

WMAX

Maximum observed individual weight (numeric).

LMIN

Minimum observed length (numeric, in mm).

LMIN01

1st percentile of length (numeric, in mm).

LMIN05

5th percentile of length (numeric, in mm).

LMIN10

10th percentile of length (numeric, in mm).

LMIN25

25th percentile of length (numeric, in mm).

LMAX75

75th percentile of length (numeric, in mm).

LMAN90

90th percentile of length calculated via mean or other method (numeric, in mm).

LMAX95

95th percentile of length (numeric, in mm).

LMAX99

99th percentile of length (numeric, in mm).

LMAX

Maximum observed length (numeric, in mm).

obs_in_TC

Number of observations found in TC data (numeric).

FAUNISTIC_CATEGORY

Old faunistic categories (character).

START_YEAR

Year when the species started to be considered target (numeric).

END_YEAR

Year when the species stopped being considered target (numeric).

GROUP

Species group or classification (character).

Source

Literature and other data sources.

Examples

data(DataTargetSpecies)
head(DataTargetSpecies)
#>   SPECIES WMIN WMIN05 WMIN10 WMIN25 WMAX_75 WMAX90 WMAX95   WMAX LMIN LMIN01
#> 1 ABRAVER 0.27   2.00   2.40   3.00    5.33   7.50  10.00 142.86   15     20
#> 2 ABRIMOR 1.00   1.30   1.60   2.50    7.54   7.62   7.64   7.67   NA     NA
#> 3 ABRRALB 0.50   0.58   0.65   0.85    1.00   1.00   1.00   1.00   NA     NA
#> 4 ABRRLON 0.33   0.38   0.43   0.50    1.00   1.00   1.00   1.00   NA     NA
#> 5 ABRRPRI 1.50   1.50   1.50   1.50    1.50   1.50   1.50   1.50   NA     NA
#> 6 ACANEXI 5.00   5.95   6.13   6.83   12.62  16.74  18.00  18.00   NA     NA
#>   LMIN05 LMIN10 LMIN25 LMAX75 LMAN90 LMAX95 LMAX99 LMAX obs_in_TC
#> 1     20     25     30     40     45     45     50   55       829
#> 2     NA     NA     NA     NA     NA     NA     NA   NA        NA
#> 3     NA     NA     NA     NA     NA     NA     NA   NA        NA
#> 4     NA     NA     NA     NA     NA     NA     NA   NA        NA
#> 5     NA     NA     NA     NA     NA     NA     NA   NA        NA
#> 6     NA     NA     NA     NA     NA     NA     NA   NA        NA
#>   FAUNISTIC_CATEGORY START_YEAR END_YEAR GROUP
#> 1                  C         NA       NA      
#> 2                  C         NA       NA      
#> 3                Emb         NA       NA      
#> 4                Emb         NA       NA      
#> 5                Emb         NA       NA      
#> 6                  B         NA       NA