R code to perform multiple checks on MEDITS Survey data (TA, TB, TC, TE and TL files) - continuous checks

The function calls all the functions included in the package in an ordered way to perform a complete quality check of TX data available. The check is performed simultaneously on the files that can contain also data of more than one year. Unlike the RoME function, RoMEcc does not stop at the first detected error, allowing user to correct data. Instead, it checks all the data and returns a report on the errors found, as well as compiling a detailed log file.

RoMEcc(TA,TB,TC,TE=NA,TL=NA,wd,suffix=NA,
verbose =TRUE,Stratification=RoME::stratification_scheme,
Ref_list=RoME::TM_list,DataTargetSpecies=RoME::DataTargetSpecies,
Maturity=RoME::Maturity_parameters,
ab_parameters=RoME::LW,
stages_list=RoME::mat_stages,assTL=RoME::assTL, zip=TRUE)

Arguments

TA

Haul data table according to MEDITS protocol (TA)

TB

Catch data table according to MEDITS protocol (TB)

TC

Biological data table according to MEDITS protocol (TC)

TE

Individual biological data table according to MEDITS protocol (TE) if available, if TE data are not available, use NA.

TL

Litter data table according to MEDITS protocol (TL) if available, if TL data are not available, use NA

wd

working directory path defined by the user

suffix

Suffix string of the Logfile, in case it suffix=NA it is automatically generated by the function using date and time of the check

verbose

...

Stratification

Stratification scheme according to MEDITS protocol.

Ref_list

TM_list reference list

DataTargetSpecies

Information related to target species.

Maturity

Information related to sex and maturity from literature or other sources.

ab_parameters

dataframe containing the a and b parameters of the length-weight relationships

stages_list

Table of maturity stages.

assTL

data frame with the association between TL (litter table) categories and sub-categories

zip

boolean, if TRUE a zip file containing the results is generated

Details

RoMEcc checks can be used to integrate a list of common quality checks on survey data. This function calls all the functions built in the package in an ordered way to perform a complete quality check of TX data available. The order of the checks in RoME was implemented in a defined sequence to avoid cascade errors due to the correction of a previous error. No automatic correction is implemented in 'RoME'. 'RoME' stops if an error occurs; then the user has to correct the error and run again the code to continue with the other checks. The function runs on a complete time series dataset, checking year after year, until the end of the time series. After the checks of the mandatory fields and the controlled vocabulary, that are carried out for all the TX tables, the specific checks on each kind of TX table are performed. Finally, RoME provides a list of cross checks aimed to guarantee the consistency among the data tables.

Some functions included in the 'RoME' library and used by RoME function need specific dictionaries or tables. It is the case of Stratification, TM_list, DataTargetSpecies, Maturity_parameters, mat_stages and assTL tables. All of them are provided by default in this library. Anyway, the user has the possibility to provide ad hoc modified versions of these tables adapting the checks to specific needs.

Value

The function does not correct data, but it detects the errors, warning the user that there is the possibility of one or more errors, specifying the type of the error and easing the data correction. If parameter verbose=TRUE returns a series of text output in console to let the user to trace the state of the checks. All the output of the functions are stored in the user defined working directory wd and in the sub-directory there resident. In the Lofile subfolder are stored the logfiles of each run of the function.

References

Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp. https://www.sibm.it/MEDITS 2011/principaledownload.htm

Author

W. Zupa,I. Bitetto

Examples

# \donttest{
wd <- tempdir()
suffix=NA
DataTA = data.frame(RoME::TA[RoME::TA$YEAR==2012 ,])
DataTB = data.frame(RoME::TB[RoME::TB$YEAR==2012 ,])
DataTC = data.frame(RoME::TC[RoME::TC$YEAR==2012 ,])
DataTE = NA
DataTL = NA
RoMEcc(DataTA, DataTB,DataTC,DataTE,DataTL, wd, suffix,
  Stratification=RoME::stratification_scheme,
  Ref_list=RoME::TM_list,DataTargetSpecies=RoME::DataTargetSpecies,
  Maturity=RoME::Maturity_parameters,ab_parameters=RoME::LW,
  stages_list=RoME::mat_stages,assTL=RoME::assTL, verbose=FALSE)
#> 
#> ########################
#> ###       2012       ###
#> ########################
#>     errors
#> 1         
#> 2         
#> 3         
#> 4         
#> 5         
#> 6         
#> 7         
#> 8         
#> 9         
#> 10        
#> 11        
#> 12        
#> 13        
#> 14        
#> 15        
#> 16        
#> 17        
#> 18        
#> 19        
#> 20        
#> 21        
#> 22        
#> 23        
#> 24        
#> 25        
#> 26        
#> 27        
#> 28        
#> 29        
#> 30        
#> 31        
#> 32        
#> 33        
#> 34        
#> 35        
#> 36        
#> 37        
#> 38        
#> 39        
#> 40        
#> 41        
#> 42        
#> 43        
#> 44        
#> 45        
#> 46        
#> 47        
#> 48        
#> 49        
#> 50        
#> 51        
#> 52        
#> 53        
#> 54        
#> 55        
#> 56        
#> 57        
#> 58        
#> 59        
#> 60        
#> 61        
#> 62        
#> 63        
#> 64        
#> 65        
#> 66        
#> 67        
#> 68        
#> 69        
#> 70        
#> 71        
#> 72        
#> 73        
#> 74        
#> 75        
#> 76        
#> 77        
#> 78        
#> 79        
#> 80        
#> 81        
#> 82        
#> 83        
#> 84        
#> 85        
#> 86        
#> 87        
#> 88        
#> 89        
#> 90        
#> 91        
#> 92        
#> 93        
#> 94        
#> 95        
#> 96        
#> 97        
#> 98        
#> 99        
#> 100       
#> 101       
#> 102       
#> 103       
#> 104       
#> 105       
#> 106       
#> 107       
#> 108       
#> 109       
#> 110       
#> 111       
#> 112       
#> 113       
#> 114       
#> 115       
#> 116       
#> 117       
#> 118       
#> 119       
#> 120       
#> 121       
#> 122       
#>                                                                                                                            check type
#> 1                                                                                                                       Check Headers
#> 2                                                                                                                       Check Headers
#> 3                                                                                                                       Check Headers
#> 4                                                                                                                         Check CLASS
#> 5                                                                                                                         Check CLASS
#> 6                                                                                                                         Check CLASS
#> 7                                                                                                                          Check YEAR
#> 8                                                                                                                  Check TYPE_OF_FILE
#> 9                                                                                                           Check identical record TA
#> 10                                                                                                          Check identical record TB
#> 11                                                                                                          Check identical record TC
#> 12                                                                                                 Check quasi-identical record in TA
#> 13                                                                                                 Check quasi-identical record in TB
#> 14                                                                                                 Check quasi-identical record in TC
#> 15                                                                                       Check consistency of area TA, TB, TC, TE, TL
#> 16                                                                                                                   Check dictionary
#> 17                                                                                                                   Check dictionary
#> 18                                                                                                                   Check dictionary
#> 19                                                                                                                   Check dictionary
#> 20                                                                                                                   Check dictionary
#> 21                                                                                                                   Check dictionary
#> 22                                                                                                                   Check dictionary
#> 23                                                                                                                   Check dictionary
#> 24                                                                                                                   Check dictionary
#> 25                                                                                                                   Check dictionary
#> 26                                                                                                                   Check dictionary
#> 27                                                                                                                   Check dictionary
#> 28                                                                                                                Check numeric range
#> 29                                                                                                                Check numeric range
#> 30                                                                                                                   Check dictionary
#> 31                                                                                                                   Check dictionary
#> 32                                                                                                                Check numeric range
#> 33                                                                                                                Check numeric range
#> 34                                                                                                                   Check dictionary
#> 35                                                                                                                   Check dictionary
#> 36                                                                                                                   Check dictionary
#> 37                                                                                                                   Check dictionary
#> 38                                                                                                                   Check dictionary
#> 39                                                                                                                   Check dictionary
#> 40                                                                                                                   Check dictionary
#> 41                                                                                                                   Check dictionary
#> 42                                                                                                                   Check dictionary
#> 43                                                                                                                   Check dictionary
#> 44                                                                                                                   Check dictionary
#> 45                                                                                                                   Check dictionary
#> 46                                                                                                                   Check dictionary
#> 47                                                                                                                   Check dictionary
#> 48                                                                                                                   Check dictionary
#> 49                                                                                                                Check numeric range
#> 50                                                                                                                Check numeric range
#> 51                                                                                                                Check numeric range
#> 52                                                                                                                Check numeric range
#> 53                                                                                                                   Check dictionary
#> 54                                                                                                                   Check dictionary
#> 55                                                                                                                   Check dictionary
#> 56                                                                                                                   Check dictionary
#> 57                                                                                                                   Check dictionary
#> 58                                                                                                                   Check dictionary
#> 59                                                                                                                   Check dictionary
#> 60                                                                                                                Check numeric range
#> 61                                                                                                                Check numeric range
#> 62                                                                                                                Check numeric range
#> 63                                                                                                                Check numeric range
#> 64                                                                                                                Check numeric range
#> 65                                                                                                                   Check dictionary
#> 66                                                                                                                   Check dictionary
#> 67                                                                                                                   Check dictionary
#> 68                                                                                                                   Check dictionary
#> 69                                                                                                                   Check dictionary
#> 70                                                                                                                   Check dictionary
#> 71                                                                                                                   Check dictionary
#> 72                                                                                                                   Check dictionary
#> 73                                                                                                                Check numeric range
#> 74                                                                                                                Check numeric range
#> 75                                                                                                                Check numeric range
#> 76                                                                                                                Check numeric range
#> 77                                                                                                                   Check dictionary
#> 78                                                                                                                   Check dictionary
#> 79                                                                                                                   Check dictionary
#> 80                                                                                                           Check no empty fields TA
#> 81                                                                                                           Check no empty fields TB
#> 82                                                                                                           Check no empty fields TC
#> 83                                                                                                                  Check 0 fields TA
#> 84                                                                                                                        Check dm TA
#> 85                                                                                     Check consistency between duration and time TA
#> 86                                                                     Check consistency between distance and duration of the haul TA
#> 87                                     Check consistency of the hauls coordinates with the distance (difference not greater than 30%)
#> 88                                                                                             Check consistency of bridles length TA
#> 89                                                    Check difference between start depth and end depth (not greater than 20%) in TA
#> 90                                                                             Check start depth and end depth in the same stratum TA
#> 91                                                                                            Check consistency of stratum code in TA
#> 92                                                                                                                   Check swept area
#> 93                                                                                           Check start quadrant and end quadrant TA
#> 94                                                                                                 Check uniqueness of valid hauls TA
#> 95                                                                                                 Visual check of the haul positions
#> 96                                          Relation between shooting depth and warp length, and between warp length and wing opening
#> 97                                                                     Check if the coordinates are in the Mediterranean or Black Sea
#> 98                                                                                                       Check on temperature by haul
#> 99                                                                                              Check correctness of species codes TB
#> 100                                                                               Check consistency of NB_TOTAL and number per sex TB
#> 101                                                                   Check presence of NB_TOTAL and number per sex TB for species G1
#> 102                                                               Check consistency between not null weight and not null total number
#> 103                                                                                         Check consistency of weight and number TB
#> 104                                                                                             Check correctness of species codes TC
#> 105                                                                                       Check correctness of LENGTH_CLASSES_CODE TC
#> 106                                                                                            Check consistency of length classes TC
#> 107                                                                                         Check correctness of number per sex in TC
#> 108                                                                                           Check consistency of maturity stages TC
#> 109                                                                                                                Check sub-sampling
#> 110 Check consistency of maturity stages TC by the comparison with the length of smallest mature individuals reported in bibliography
#> 111                                                                                       Check consistency of length distribution TC
#> 112                                                                                              Check total weight in the haul in TC
#> 113                                                                                             Check correctness of species codes TC
#> 114                                                                      Check presence of lengths for G1 and G2 Medits species in TC
#> 115                                                                                                  Check presence in TB of TA hauls
#> 116                                                                                                  Check presence in TA of TB hauls
#> 117                                                                                                                   Check abundance
#> 118                                                                                         Check presence in TC of TB target species
#> 119                                                                                                Check presence in TB of TC species
#> 120                                                       Check correctness of the number per sex in TB in case of sub-sampling in TC
#> 121                                                                                                          Check on date by haul TB
#> 122                                                                                                          Check on date by haul TC
#>     table                                                     variable year
#> 1      TA                                                         <NA>   NA
#> 2      TB                                                         <NA>   NA
#> 3      TC                                                         <NA>   NA
#> 4      TA                                                         <NA>   NA
#> 5      TB                                                         <NA>   NA
#> 6      TC                                                         <NA>   NA
#> 7     ALL                                                         <NA>   NA
#> 8     ALL                                                         <NA>   NA
#> 9      TA                                                         <NA> 2012
#> 10     TB                                                         <NA> 2012
#> 11     TC                                                         <NA> 2012
#> 12     TA                                                         <NA> 2012
#> 13     TB                                                         <NA> 2012
#> 14     TC                                                         <NA> 2012
#> 15    ALL                                                         <NA> 2012
#> 16     TA                                                      COUNTRY 2012
#> 17     TA                                                         AREA 2012
#> 18     TA                                                         GEAR 2012
#> 19     TA                                                      RIGGING 2012
#> 20     TA                                                        DOORS 2012
#> 21     TA                                                        MONTH 2012
#> 22     TA                                                          DAY 2012
#> 23     TA                                                  HAUL_NUMBER 2012
#> 24     TA                                               CODEND_CLOSING 2012
#> 25     TA                                           PART_OF_THE_CODEND 2012
#> 26     TA                                                SHOOTING_TIME 2012
#> 27     TA                                            SHOOTING_QUADRANT 2012
#> 28     TA                                            SHOOTING_LATITUDE 2012
#> 29     TA                                           SHOOTING_LONGITUDE 2012
#> 30     TA                                               SHOOTING_DEPTH 2012
#> 31     TA                                             HAULING_QUADRANT 2012
#> 32     TA                                             HAULING_LATITUDE 2012
#> 33     TA                                            HAULING_LONGITUDE 2012
#> 34     TA                                                 HAULING_TIME 2012
#> 35     TA                                                HAULING_DEPTH 2012
#> 36     TA                                                HAUL_DURATION 2012
#> 37     TA                                                     VALIDITY 2012
#> 38     TA                                                       COURSE 2012
#> 39     TA                                                     DISTANCE 2012
#> 40     TA                                             VERTICAL_OPENING 2012
#> 41     TA                                                 WING_OPENING 2012
#> 42     TA                                        GEOMETRICAL_PRECISION 2012
#> 43     TA                                             MEASURING_SYSTEM 2012
#> 44     TA                                    MEASURING_SYSTEM_SALINITY 2012
#> 45     TA                                             RECORDED_SPECIES 2012
#> 46     TA                                                  WARP_LENGTH 2012
#> 47     TA                                                WARP_DIAMETER 2012
#> 48     TA                                                 OBSERVATIONS 2012
#> 49     TA                                 BOTTOM_TEMPERATURE_BEGINNING 2012
#> 50     TA                                       BOTTOM_TEMPERATURE_END 2012
#> 51     TA                                    BOTTOM_SALINITY_BEGINNING 2012
#> 52     TA                                          BOTTOM_SALINITY_END 2012
#> 53     TB                                                      COUNTRY 2012
#> 54     TB                                                         AREA 2012
#> 55     TB                                                        MONTH 2012
#> 56     TB                                                          DAY 2012
#> 57     TB                                                  HAUL_NUMBER 2012
#> 58     TB                                               CODEND_CLOSING 2012
#> 59     TB                                           PART_OF_THE_CODEND 2012
#> 60     TB                                     TOTAL_WEIGHT_IN_THE_HAUL 2012
#> 61     TB                                     TOTAL_NUMBER_IN_THE_HAUL 2012
#> 62     TB                                                NB_OF_FEMALES 2012
#> 63     TB                                                  NB_OF_MALES 2012
#> 64     TB                                           NB_OF_UNDETERMINED 2012
#> 65     TC                                                      COUNTRY 2012
#> 66     TC                                                         AREA 2012
#> 67     TC                                                        MONTH 2012
#> 68     TC                                                          DAY 2012
#> 69     TC                                                  HAUL_NUMBER 2012
#> 70     TC                                               CODEND_CLOSING 2012
#> 71     TC                                           PART_OF_THE_CODEND 2012
#> 72     TC                                          LENGTH_CLASSES_CODE 2012
#> 73     TC                                       WEIGHT_OF_THE_FRACTION 2012
#> 74     TC                                WEIGHT_OF_THE_SAMPLE_MEASURED 2012
#> 75     TC                   NO_OF_INDIVIDUAL_OF_THE_ABOVE_SEX_MEASURED 2012
#> 76     TC NUMBER_OF_INDIVIDUALS_IN_THE_LENGTH_CLASS_AND_MATURITY_STAGE 2012
#> 77     TC                                                          SEX 2012
#> 78     TC                                                 LENGTH_CLASS 2012
#> 79     TC                                                     MATURITY 2012
#> 80     TA                                                         <NA> 2012
#> 81     TB                                                         <NA> 2012
#> 82     TC                                                         <NA> 2012
#> 83     TA                                                         <NA> 2012
#> 84     TA                                                         <NA> 2012
#> 85     TA                                                         <NA> 2012
#> 86     TA                                                         <NA> 2012
#> 87     TA                                                         <NA> 2012
#> 88     TA                                                         <NA> 2012
#> 89     TA                                                         <NA> 2012
#> 90     TA                                                         <NA> 2012
#> 91     TA                                                         <NA> 2012
#> 92     TA                                                         <NA> 2012
#> 93     TA                                                         <NA> 2012
#> 94     TA                                                         <NA> 2012
#> 95     TA                                                         <NA> 2012
#> 96     TA                                                         <NA> 2012
#> 97     TA                                                         <NA> 2012
#> 98     TA                                                         <NA> 2012
#> 99     TB                                                         <NA> 2012
#> 100    TB                                                         <NA> 2012
#> 101    TB                                                         <NA> 2012
#> 102    TB                                                         <NA> 2012
#> 103    TB                                                         <NA> 2012
#> 104    TC                                                         <NA> 2012
#> 105    TC                                                         <NA> 2012
#> 106    TC                                                         <NA> 2012
#> 107    TC                                                         <NA> 2012
#> 108    TC                                                         <NA> 2012
#> 109    TC                                                         <NA> 2012
#> 110    TC                                                         <NA> 2012
#> 111    TC                                                         <NA> 2012
#> 112    TC                                                         <NA> 2012
#> 113    TC                                                         <NA> 2012
#> 114    TC                                                         <NA> 2012
#> 115 TB-TA                                                         <NA> 2012
#> 116 TA-TB                                                         <NA> 2012
#> 117 TA-TB                                                         <NA> 2012
#> 118 TC-TB                                                         <NA> 2012
#> 119 TB-TC                                                         <NA> 2012
#> 120 TB-TC                                                         <NA> 2012
#> 121 TA-TB                                                         <NA> 2012
#> 122 TA-TC                                                         <NA> 2012
# }