TM list

The present list is destined to code the marine species encountered in the Mediterranean. It has been built following the principle used in the Nordic Code Centre (Stockholm). For most of the species the codes are identical to those proposed by the NCC. However some species can be coded differently. In addition numerous Mediterranean species are not included in the NCC code and have been added. So the present list is specific. It has to be referred as the TM list (Taxonomic list not only Faunistic, FM list).

data("TM_list")

Format

A data frame with 1470 observations on the following 11 variables.

N.

a numeric vector

MeditsCode

a factor with levels of species codes in the RUBIN format (see MEDITS manual)

Scientific.Name.................................................valid

a factor with levels of the scientific names of the species

Authorship

a factor with levels of the authorship of the inforformation

Source

a factor with levels sources of the inforformation

Reference

a factor with levels of the bibliografic references

Remarks

a factor with levels the reported remarks

CATFAU

a factor with levels of the faunistic categories of the species

CODLON

a factor with CODLON that represents the Length classes code: m = 1 mm; 0 = 0,5 cm; 1 = 1 cm.

GSAs

a factor with levels of the Geographic Sub-Areas (GSA) adopted in the MEDITS protocol.

Year

a factor with levels of the years

Source

MEDITS MEDITS-Handbook, Version n. 9 (2017)

References

Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp. http://www.sibm.it/MEDITS 2011/principaledownload.htm

Examples

data(TM_list)
str(TM_list)
#> 'data.frame':	1617 obs. of  11 variables:
#>  $ N.                                                                   : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ MeditsCode                                                           : chr  "ABRAVER" "ABRIMOR" "ABRRALB" "ABRRLON" ...
#>  $ Scientific.Name.................................................valid: chr  "Abralia veranyi" "Abraliopsis morisii" "Abra alba" "Abra longicallus" ...
#>  $ Authorship                                                           : chr  "(Rüppell, 1844)" "(Vérany, 1839)" "(Wood, 1802)" "(Scacchi, 1835)" ...
#>  $ Source                                                               : chr  "FP" "FP" "R" "" ...
#>  $ Reference                                                            : chr  "ENOP, 121" "ENOP, 123s" "p. 339 Tav. 129" "" ...
#>  $ Remarks                                                              : chr  "" "" "" "" ...
#>  $ CATFAU                                                               : chr  "C" "C" "Emb" "Emb" ...
#>  $ CODLON                                                               : chr  "0" "0" "0" "" ...
#>  $ GSAs                                                                 : chr  "1, 2, 5, 6, 7, 8, 9, 10, 11, 15, 16, 17, 18, 19, 20, 22, 23, 25" "19" "1, 6, 16, 19" "1, 19" ...
#>  $ Year                                                                 : chr  "" "2011 (19)" "" "2013 (1)" ...
head(TM_list)
#>   N. MeditsCode
#> 1  1    ABRAVER
#> 2  2    ABRIMOR
#> 3  3    ABRRALB
#> 4  4    ABRRLON
#> 5  5    ABRRPRI
#> 6  6    ACANEXI
#>   Scientific.Name.................................................valid
#> 1                                                       Abralia veranyi
#> 2                                                   Abraliopsis morisii
#> 3                                                             Abra alba
#> 4                                                      Abra longicallus
#> 5                                                       Abra prismatica
#> 6                                                   Acanthephyra eximia
#>        Authorship Source       Reference Remarks CATFAU CODLON
#> 1 (Rüppell, 1844)     FP       ENOP, 121              C      0
#> 2  (Vérany, 1839)     FP      ENOP, 123s              C      0
#> 3    (Wood, 1802)      R p. 339 Tav. 129            Emb      0
#> 4 (Scacchi, 1835)                                   Emb       
#> 5 (Montagu, 1808)                                   Emb      0
#> 6     Smith, 1884      Z              84              B      m
#>                                                              GSAs      Year
#> 1 1, 2, 5, 6, 7, 8, 9, 10, 11, 15, 16, 17, 18, 19, 20, 22, 23, 25          
#> 2                                                              19 2011 (19)
#> 3                                                    1, 6, 16, 19          
#> 4                                                           1, 19  2013 (1)
#> 5                                                            7, 8      2017
#> 6                                    1, 2, 5, 6, 7, 9, 11, 16, 19