Function to create R-sufi file containing haul data.

This function reports the information contained in the hauld data table (TA) from the MEDITS protocol to the format required by R-sufi (Rochet et al., 2004).

create_haul(ResultDataTA,year,wd,save=TRUE)

Arguments

ResultDataTA

Haul data table according to MEDITS protocol (TA)

year

reference year for the analysis

wd

working directory path defined by the user

save

boolean value to save the results in a csv file

Value

The function saves by default in the files R-Sufi folder the table traits in.csv format, with suffix of the year and GSA. If save parameter is FALSE the function returns the data frame as output.

References

Rochet M. J., V. M. Trenkel, J. A. Bertrand & J.-C. Poulard, 2004. R routines for survey based fisheries population and community indicators (R-SUFI). Ifremer, Nantes. Limited distribution. Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp.https://www.sibm.it/MEDITS 2011/principaledownload.htm

Author

I. Bitetto, W. Zupa

Examples

TA = RoME::TA
year = 2012
wd = tempdir()
create_haul(TA,year,wd,save=FALSE)
#>          Survey Year Haul Month Stratum SweptSurface      Lat     Long Depth
#> 1  MEDITS-GSA10 2012    1     6       4          0.1 40.94667 13.68158 330.5
#> 2  MEDITS-GSA10 2012    2     6       3        0.046   41.016 13.70192   111
#> 3  MEDITS-GSA10 2012    3     6       3        0.046 40.90833 13.83333   110
#> 4  MEDITS-GSA10 2012    4     6       2        0.045 40.95675   13.791    94
#> 5  MEDITS-GSA10 2012    5     6       2        0.044 40.83067 13.96025  74.5
#> 6  MEDITS-GSA10 2012    6     6       1        0.043 40.91758 13.93417    24
#> 7  MEDITS-GSA10 2012    7     6       5        0.099   40.526   13.931 532.5
#> 8  MEDITS-GSA10 2012    8     6       4        0.097 40.61475 14.00958 462.5
#> 9  MEDITS-GSA10 2012   10     6       5          0.1 40.43317 14.29175 676.5
#> 10 MEDITS-GSA10 2012   21     6       1        0.042 40.45742 14.86875  43.5