Estimation of haul distance

Function to estimate the hauls length using TA (table A, hauls data) with coordinates in the decimal degrees format (dd.ddd). The distances could be returned expressed in meters, kilometers and nautical miles.

dd.distance(data, unit = "m", verbose=TRUE)

Arguments

data

data frame of the hauls data (TA, table A)

unit

string value indicating the measure unit of the distance. Allowed values: "m" for meters, "km" for kilometers and "NM" for nautical miles.

verbose

give verbose output reporting in the output the selected measure unit of the distance.

Details

The TA file should be populated with coordinates in dicimal degrees format.

Value

The function returns the vector of the distances expressed in the selected measure unit.

Author

Walter Zupa

Examples

TA.dd <- MEDITS.to.dd(TA)
dd.distance(TA.dd, unit="km", verbose=FALSE)
#>   [1] 5.409739 2.531317 3.403627 2.454159 2.634717 2.271499 4.899833 5.524425
#>   [9] 5.429025 2.782729 5.214030 2.686961 2.612837 2.851746 2.801794 2.804268
#>  [17] 4.831533 5.098002 5.053704 2.755603 5.104576 2.631401 2.694840 2.753682
#>  [25] 2.450797 2.612513 4.893180 5.013189 4.810696 2.494795 5.396158 2.666596
#>  [33] 2.720887 2.517513 2.603268 2.630999 4.995205 5.144671 5.125725 2.794506
#>  [41] 4.995738 2.647977 2.686012 2.894716 2.568665 2.645541 5.071301 4.949117
#>  [49] 4.761716 2.500096 5.243339 2.598423 2.575637 2.597429 2.596458 3.337424
#>  [57] 5.120894 5.040618 4.186102 2.624323 5.328334 2.749170 2.593118 2.809438
#>  [65] 2.818290 2.735397 5.249962 5.057563 4.974280 2.522611 5.258758 2.673755
#>  [73] 2.683679 2.602219 2.497751 2.699203 5.110578 5.053500 4.591926 2.771086
#>  [81] 6.834721 2.710136 2.757628 2.769897 2.756772 2.717418 5.151445 5.216785
#>  [89] 4.900522 3.246467 2.801000 2.716307 2.382340 2.916204 2.755633 2.791263
#>  [97] 5.467045 5.304127 5.215077 5.083646